Secant, Fixed, and BiSection Modifications :smile:
General Modifications:
1- Adding comments to all methods.
2-Modifing the name of the file.
-----------------------------------------------------------------------
Secant Modifications 
There is nothing happend in this method, I searched for bounds said today, but i found nothing about it in the slides of the doctor, If you found any please notify me and I will do it at once.
------------------------------------------------------------------------
Fixed Method Modifications
Change the returned array to be 2 parts, First part for  specifications of each Iteration, and the second part is the err part to indicates if the method will converge or divarge, if err = 1 the method divarge.
-------------------------------------------------------------------------
BiSection Method modifications
I added a new element to the array returned from this method which is the number of Iterations theoritcally as explaind in the slide whose number is 9 and located in presentation named Lec_6.
----------------------------------------------------------------------------
General Note.
If anyone have any comment please notify me inorder to change it at once.